+++ packages/workbox-streams/src/isSupported.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import './_version.mjs';
+import './_version.js';
 
 let cachedIsSupported = undefined;
