+++ packages/workbox-streams/src/utils/createHeaders.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import '../_version.mjs';
+import '../_version.js';
 
 /**
  * This is a utility method that determines whether the current browser supports
